Restructure storybook a bit
