--- packages/components/src/components/editor/editor.css.ts
@@ -1,5 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const editor = style({
-  position: 'relative'
-});
