+++ packages/components/src/components/editor/editor.stories.tsx
@@ -1,48 +1,44 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Editor } from './editor';
-import { Default as ToolbarStory } from '../master/toolbar/toolbar.stories';
+import { Default as ToolbarStory } from './toolbar/toolbar.stories';
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup, Flex } from '@/components/common';
-import { SidebarHeader } from '../detail/header/header';
+import { SidebarHeader } from './sidebar/header';
 import { IvyIcons } from '@axonivy/ui-icons';
 import * as React from 'react';
 
-const meta: Meta<typeof Editor> = {
-  title: 'Editor',
-  component: Editor
+const meta: Meta = {
+  title: 'Editor'
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Editor>;
+type Story = StoryObj;
 
 const EditorStory = () => {
   const [sideBar, setSideBar] = React.useState(true);
   return (
-    <Editor>
-      <ResizablePanelGroup direction='horizontal' style={{ minHeight: 200 }}>
-        <ResizablePanel defaultSize={75} minSize={50}>
-          <Flex direction='column' style={{ height: '100%' }}>
-            <ToolbarStory sideBarCollapse={() => setSideBar(old => !old)} />
-            <Flex justifyContent='center' alignItems='center' style={{ height: '100%' }}>
-              <span>Content</span>
-            </Flex>
+    <ResizablePanelGroup direction='horizontal' style={{ minHeight: 200 }}>
+      <ResizablePanel defaultSize={75} minSize={50}>
+        <Flex direction='column' style={{ height: '100%' }}>
+          <ToolbarStory sideBarCollapse={() => setSideBar(old => !old)} />
+          <Flex justifyContent='center' alignItems='center' style={{ height: '100%' }}>
+            <span>Content</span>
           </Flex>
-        </ResizablePanel>
-        {sideBar && (
-          <>
-            <ResizableHandle />
-            <ResizablePanel defaultSize={25} minSize={10}>
-              <Flex direction='column' style={{ height: '100%' }}>
-                <SidebarHeader icon={IvyIcons.Script} title='Script' />
-                <Flex justifyContent='center' alignItems='center' style={{ height: '100%' }}>
-                  <span>Sidebar</span>
-                </Flex>
+        </Flex>
+      </ResizablePanel>
+      {sideBar && (
+        <>
+          <ResizableHandle />
+          <ResizablePanel defaultSize={25} minSize={10}>
+            <Flex direction='column' style={{ height: '100%' }}>
+              <SidebarHeader icon={IvyIcons.Script} title='Script' />
+              <Flex justifyContent='center' alignItems='center' style={{ height: '100%' }}>
+                <span>Sidebar</span>
               </Flex>
-            </ResizablePanel>
-          </>
-        )}
-      </ResizablePanelGroup>
-    </Editor>
+            </Flex>
+          </ResizablePanel>
+        </>
+      )}
+    </ResizablePanelGroup>
   );
 };
