+++ packages/components/src/components/editor/index.ts
@@ -1,0 +1,2 @@
+export * from './sidebar/header';
+export * from './toolbar/toolbar';
