+++ packages/components/src/components/editor/quickbar/quickbar.tsx
