+++ packages/components/src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './common';
-export * from './detail/header/header';
-export * from './master/toolbar/toolbar';
+export * from './editor';
